Fixed some bugs
Particularly, I changed it so the responsive alignment no longer changes
the object alignment settings, only temporary for current display.
Also fixed a bug where the arrow positioning wasn't being reset, this was
causing multiple arrows to appear if you were to open a popup, then
scroll, and have it reopen in the flipped position.

I also removed the iFrame support, as it won't work cross-browser (i only
really put it in there as test anyhow)