Changed popup positioning
Now used the popup-container's margin as the buffer zone for the arrow.
when the arrow is set to be shown, css should add a margin to the side
that the arrow will be placed
this will be used to calculate the extra space needed for the arrow.
the "buffer" variable has been completely removed, and is handing over
complete control of the arrow buffering zone to the css