Added a new animation test called Flip
It uses Transform: RotateX and SkewX, which aren't support in webkit, even
with -webkit-transform. This was to test how my getTotalTransitionTime
method would deal with none supported css when figuring out the transition
time. As I expected it doesn't actually get added to the elements style so
my method returns "0s" if it's the only transition there