added bower.json for future bower release
i'm not using bower to bring in any devDependencies. I find it weird that
bower is pulling in the entire repo for jquery and bootstrap. I wanted to
use it to bull in just the min version of the two for the demo, but i'm
just going to stay with the copies i currently have in demo/assets/ for
now